[FE-add-photo-to-form-motorcycle] fix lead remark, change myself request hook on useQuery in catalog page
