+++ frontend/src/pages/CatalogMotorcycles/index.tsx
@@ -1,107 +1,40 @@
-import { useEffect, useState } from 'react';
-import { useHttp } from '../../hooks/http.hook';
+import { useQuery } from 'react-query';
+
 import { API_V1_URL } from '../../constants';
+import { motorcycleData } from './types';
 import MotorcycleCard, {
   IMotorcycleData,
 } from '../../components/MotorcycleCard';
 
-export interface IMotorcycleCard {
-  name: string;
-  price: number;
-  vendorCode: string;
-  description: string;
-}
-
-export interface IError {
-  errorCode: number;
-  message: string;
-}
-
-const motorcycleData: IMotorcycleData[] = [
-  {
-    name: 'LONCIN LX200GY-3 PRUSS',
-    price: 61760,
-    vendorCode: 'MOT-225',
-    availableColor: ['red', 'black', 'yellow', 'green'],
-    cubicCapacity: 150,
-    maxSpeed: 115,
-    numberOfGears: 5,
-    brakes: 'Дискові/Барабанні',
-    fuelInjection: 'Карбюратор',
-    cooling: 'Повітряне',
-  },
-  {
-    name: 'LONCIN LX200GY-3 PRUSS',
-    price: 61760,
-    vendorCode: 'MOT-225',
-    availableColor: ['red', 'black', 'yellow', 'green'],
-    cubicCapacity: 150,
-    maxSpeed: 115,
-    numberOfGears: 5,
-    brakes: 'Дискові/Барабанні',
-    fuelInjection: 'Карбюратор',
-    cooling: 'Повітряне',
-  },
-  {
-    name: 'LONCIN LX200GY-3 PRUSS',
-    price: 61760,
-    vendorCode: 'MOT-225',
-    availableColor: ['red', 'black', 'yellow', 'green'],
-    cubicCapacity: 150,
-    maxSpeed: 115,
-    numberOfGears: 5,
-    brakes: 'Дискові/Барабанні',
-    fuelInjection: 'Карбюратор',
-    cooling: 'Повітряне',
-  },
-];
+const getCatalogMotorcycle = async () => {
+  const response = await fetch(`${API_V1_URL}/motorcycleCards/allMotorcycle`, {
+    method: 'GET',
+  });
+  const formattedResponse = await response.json();
+  return formattedResponse;
+};
 
 const CatalogMotorcycles = () => {
-  const [allMotorcycles, setAllMotorcycles] = useState<IMotorcycleCard[]>([]);
-  const [errorResponse, setErrorResponse] = useState<IError | null>(null);
-  console.log('errorResponse', errorResponse);
+  const { data } = useQuery('catalogMotorcycle', {
+    queryFn: getCatalogMotorcycle,
+    onSuccess: (response) => {
+      console.log('response success', response);
+    },
+    onError: (response) => {
+      console.log('response error', response);
+    },
+  });
 
-  const { request } = useHttp();
-
-  useEffect(() => {
-    const getAllMotorcycles = async () => {
-      try {
-        const response = await request(
-          `${API_V1_URL}/motorcycleCards/allMotorcycle`,
-          'GET',
-        );
-
-        console.log('response', response);
-        console.log('response.message', response.message);
-        console.log('response.message.length', response.message.length);
-
-        if (response.errorCode == 404) {
-          return setErrorResponse(response);
-        }
-
-        setAllMotorcycles(response.message);
-      } catch (err) {
-        console.log('error', err);
-        return err;
-      }
-    };
-    getAllMotorcycles();
-  }, [request]);
-
-  useEffect(() => {
-    console.log('allMotorcycles', allMotorcycles);
-
-    console.log('API_V1_URL', JSON.stringify(API_V1_URL));
-  }, [allMotorcycles]);
+  console.log('data useQuery', data);
 
   return (
     <>
       <div>
-        {motorcycleData.map((el: IMotorcycleData) => {
+        {motorcycleData.map((el: IMotorcycleData, index: number) => {
           return (
-            <>
+            <div key={index}>
               <MotorcycleCard motorcycleData={el} />
-            </>
+            </div>
           );
         })}
       </div>
